02_build/command-centre/backend/requirements-voice.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
httpx>=0.24.0
falkordb>=1.0.0
python-multipart>=0.0.6
